+++ packages/core/src/client.ts
@@ -1,6 +1,5 @@
 import { interpret } from 'xstate'
 
-import { MIN_TOKEN_REFRESH_INTERVAL } from './constants'
 import { AuthMachine, AuthMachineOptions, createAuthMachine } from './machines'
 import { defaultClientStorageGetter, defaultClientStorageSetter } from './storage'
 import type { AuthInterpreter } from './types'
@@ -24,7 +23,7 @@ export class AuthClient {
     clientUrl = (typeof window !== 'undefined' && window.location?.origin) || '',
     clientStorageGetter = defaultClientStorageGetter,
     clientStorageSetter = defaultClientStorageSetter,
-    refreshIntervalTime = MIN_TOKEN_REFRESH_INTERVAL,
+    refreshIntervalTime,
     autoSignIn = true,
     autoRefreshToken = true,
     start = true
@@ -48,13 +47,17 @@ export class AuthClient {
     }
 
     if (typeof window !== 'undefined' && autoSignIn) {
-      this._channel = new BroadcastChannel('nhost')
-      this._channel.addEventListener('message', (token) => {
-        const existingToken = this.interpreter?.state.context.refreshToken
-        if (this.interpreter && token.data !== existingToken) {
-          this.interpreter.send({ type: 'TRY_TOKEN', token: token.data })
-        }
-      })
+      try {
+        this._channel = new BroadcastChannel('nhost')
+        this._channel.addEventListener('message', (token) => {
+          const existingToken = this.interpreter?.state.context.refreshToken
+          if (this.interpreter && token.data !== existingToken) {
+            this.interpreter.send({ type: 'TRY_TOKEN', token: token.data })
+          }
+        })
+      } catch (error) {
+        // * BroadcastChannel is not available e.g. react-native
+      }
     }
   }