+++ packages/apollo/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @nhost/apollo
 
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies [5ee395e]
+- Updated dependencies [e0cfcaf]
+- Updated dependencies [7b7527a]
+  - @nhost/core@0.3.13
+
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [7b5f00d]
+- Updated dependencies [58e1485]
+  - @nhost/core@0.3.12
+
 ## 0.4.1
 
 ### Patch Changes