+++ packages/react-apollo/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @nhost/react-apollo
 
+## 4.0.17
+
+### Patch Changes
+
+- @nhost/react@0.5.7
+- @nhost/apollo@0.4.3
+
+## 4.0.16
+
+### Patch Changes
+
+- Updated dependencies [5ee395e]
+  - @nhost/react@0.5.6
+  - @nhost/apollo@0.4.3
+
+## 4.0.15
+
+### Patch Changes
+
+- @nhost/apollo@0.4.2
+- @nhost/react@0.5.5
+
 ## 4.0.14
 
 ### Patch Changes